requirements.txt
fitz
PyMuPDF
easyocr
pandas
streamlit
python-dotenv
groq